// src/app/core/model/user.ts
/*****************************************************************************************************
 *@Purpose - FundoNotes.
 *@file    - user.ts
 *@author  - Ankita Mapari <mapariit@gmail.com>
 *@version - 1.0
 *@since   - 22/04/2019
**************************************************************************************************/
export class User {
    "id":string;
    "firstName": string;
    "lastName": string;
    "email": string;
    "password": string;
    "confirmPassword": string;
    "service": string;
}